requirements.txt
asgiref==3.8.1
Django==5.1.6
django-model-utils==5.0.0
djangorestframework==3.15.2
six==1.17.0
sqlparse==0.5.3
